backend/requirements.txt
aiofiles==0.5.0
asgiref==3.2.10
attrs==19.3.0
captcha==0.3
click==7.1.2
colorama==0.4.3
constantly==15.1.0
dnspython==2.0.0
email-validator==1.1.1
et-xmlfile==1.0.1
fastapi==0.60.0
fdfs-client-py3==1.0.0
h11==0.9.0
horetu==0.4.4
hyperlink==19.0.0
idna==2.10
incremental==17.5.0
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
MarkupSafe==1.1.1
Pillow==7.2.0
pydantic==1.6.1
PyJWT==1.7.1
PyMySQL==0.10.0
python-dotenv==0.14.0
python-multipart==0.0.5
pytz==2020.1
qrcode==6.1
redis==3.5.3
six==1.15.0
SQLAlchemy==1.3.18
sqlalchemy-pagination==0.0.2
sqlparse==0.3.1
starlette==0.13.4
uvicorn==0.11.6
WebOb==1.8.6
websockets==8.1
Werkzeug==1.0.1
XlsxWriter==1.3.3
zope.interface==5.1.0
